Moved some methods around from BaseState to their respective states
