+++ Assets/Dev/FollowingEnemy/Scripts/FollowingEnemyStates/FollowingState.cs
@@ -6,11 +6,13 @@ public class FollowingState : BaseState
     public FollowingState(FollowingEnemyScript enemy) : base(enemy) { }
     public override void Enter()
     {
+        // TODO: Remove after review
         Debug.Log("ENTERING: Following");
     }
 
     public override void Exit()
     {
+        // TODO: Remove after review
         Debug.Log("EXITING: Following");
     }
 
@@ -19,9 +21,8 @@ public class FollowingState : BaseState
         Following();
         if (enemy.target == null || !IsPlayerInSight())
         {
-            enemy.ChangeState(enemy.states["Roaming"]);
             enemy.target = null;
-            SetNewRoamingPosition();
+            enemy.ChangeState(enemy.states["Roaming"]);
         }
     }
     private void Following()
@@ -32,4 +33,26 @@ public class FollowingState : BaseState
         }
     }
 
+    // Checks if the player is still in sight
+    protected bool IsPlayerInSight()
+    {
+        if (enemy.target == null) return false;
+
+        Vector3 directionToPlayer = (enemy.target.position - enemy.transform.position).normalized;
+        float angleToPlayer = Vector3.Angle(enemy.transform.forward, directionToPlayer);
+
+        if (angleToPlayer < enemy.visionAngle / 2 && Vector3.Distance(enemy.transform.position, enemy.target.position) <= enemy.visionRange)
+        {
+            RaycastHit hit;
+            if (Physics.Raycast(enemy.transform.position, directionToPlayer, out hit, enemy.visionRange))
+            {
+                if (hit.collider.GetComponent<PlayerScript>() != null)
+                {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
 }
+
