+++ Assets/Dev/FollowingEnemy/Throwaway/FollowingEnemyScene.unity
@@ -672,7 +672,7 @@ GameObject:
   m_Component:
   - component: {fileID: 337718743}
   - component: {fileID: 337718742}
-  - component: {fileID: 337718741}
+  - component: {fileID: 337718745}
   - component: {fileID: 337718744}
   m_Layer: 0
   m_Name: CinemachineCamera
@@ -681,25 +681,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &337718741
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 337718739}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b617507da6d07e749b7efdb34e1173e1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  TrackerSettings:
-    BindingMode: 1
-    PositionDamping: {x: 1, y: 1, z: 1}
-    AngularDampingMode: 0
-    RotationDamping: {x: 1, y: 1, z: 1}
-    QuaternionDamping: 1
-  FollowOffset: {x: 0, y: 1.8, z: -2}
 --- !u!114 &337718742
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -752,7 +733,7 @@ Transform:
   m_GameObject: {fileID: 337718739}
   serializedVersion: 2
   m_LocalRotation: {x: 0.000000026617887, y: 0.000000077096324, z: 0.000000001436236, w: 1}
-  m_LocalPosition: {x: 0, y: 3.3999999, z: -3.2600002}
+  m_LocalPosition: {x: 0, y: 3, z: -2.6600003}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -770,7 +751,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 62d54399b04b4fa48a73e121d4937dd4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  ReferenceFrame: 0
+  ReferenceFrame: 3
   RecenterTarget: 0
   PanAxis:
     Value: 0
@@ -792,6 +773,32 @@ MonoBehaviour:
       Wait: 1
       Time: 2
     Restrictions: 0
+--- !u!114 &337718745
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 337718739}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 625c14eee3ad46c99df7c7c891ef668a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Damping: {x: 0.1, y: 0.5, z: 0.3}
+  ShoulderOffset: {x: 0, y: 1, z: 0}
+  VerticalArmLength: 0.4
+  CameraSide: 1
+  CameraDistance: 1.4
+  AvoidObstacles:
+    Enabled: 0
+    CollisionFilter:
+      serializedVersion: 2
+      m_Bits: 1
+    IgnoreTag: 
+    CameraRadius: 0.2
+    DampingIntoCollision: 0
+    DampingFromCollision: 0.5
 --- !u!1 &411964070
 GameObject:
   m_ObjectHideFlags: 0
@@ -924,7 +931,7 @@ Transform:
   m_GameObject: {fileID: 411964070}
   serializedVersion: 2
   m_LocalRotation: {x: 0.000000026617887, y: 0.000000077096324, z: -2.0521412e-15, w: 1}
-  m_LocalPosition: {x: 0, y: 3.3999999, z: -3.2600002}
+  m_LocalPosition: {x: 0, y: 3, z: -2.6600003}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1225,8 +1232,11 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   health: 100
+  agent: {fileID: 0}
   roamingRadius: 5
   roamingAngleRange: 90
+  roamingNextDestination: {x: 0, y: 0, z: 0}
+  target: {fileID: 0}
   followRange: 10
   visionAngle: 45
   visionRange: 10
